Sun, Jun 5, 2022, 4:31 AM +00:00

+++ readability/readability.c
@@ -18,8 +18,8 @@ int main(void)
     int sentences_amount = count_sentences(text);
     printf("%i letters\n%i words\n%i sentences\n", letters_amount, words_amount, sentences_amount);
 
-    int l = (letters_amount / words_amount) * 100;
-    printf ("l= %i\n", l);
+    float l = (letters_amount / words_amount) * 100;
+    printf ("l= %f\n", l);
     int s = (sentences_amount / words_amount) * 100;
 
     float index = 0.0588 * l - 0.296 * s - 15.8;
